Make the embedded_c sample Makefile use the main Makefile
This way it is much easier to use the same configuration flags in both
the main library code and the sample code.

Note that this also forcefully disables some of the warning types
enabled in the main file. This sample was full of warnings, but they
were never really seen.